課題2/課題2-1-(c)/requirements.txt
kafka-python==2.1.5
paho-mqtt==2.1.0
